refractor: FE OverviewPage -> move showStudentInfohandler to DiplomaCardContainer
